PHP and Javascript Confirm Before Delete
PHP and Javascript Confirm Before Delete

+++ admin/includes/view_all_posts.php
@@ -145,7 +145,7 @@
                    echo"<td>{$post_date}</td>";
                    echo"<td><a href='../post.php?p_id=$post_id'>View</a></td>"; 
                    echo"<td><a href='posts.php?source=edit_post&edit=$post_id'>Edit</a></td>"; 
-                   echo"<td><a href='posts.php?delete=$post_id'>Delete</a></td>";
+                   echo"<td><a onClick=\"javascript: return confirm('Are You Sure You Want to Delete This Message?');\" href='posts.php?delete=$post_id'>Delete</a></td>";
                    
                    
                 echo "</tr>";
